+++ crates/lib/tests/multiplication.rs
@@ -35,15 +35,15 @@ error!(
 test!(
     num_mul_nan,
     "a {\n  color: 1 * (0/0);\n}\n",
-    "a {\n  color: NaN;\n}\n"
+    "a {\n  color: calc(NaN);\n}\n"
 );
 test!(
     nan_mul_num,
     "a {\n  color: (0/0) * 1;\n}\n",
-    "a {\n  color: NaN;\n}\n"
+    "a {\n  color: calc(NaN);\n}\n"
 );
 test!(
     nan_mul_nan,
     "a {\n  color: (0/0) * (0/0);\n}\n",
-    "a {\n  color: NaN;\n}\n"
+    "a {\n  color: calc(NaN);\n}\n"
 );