+++ crates/lib/tests/special-functions.rs
@@ -45,7 +45,7 @@ error!(
 error!(
     calc_complex_unit,
     "a {\n  color: calc(1% + 1px * 2px);\n}\n",
-    "Error: Number 2px*px isn't compatible with CSS calculations."
+    "Error: Number calc(2px * 1px) isn't compatible with CSS calculations."
 );
 error!(
     calc_nested_parens,